+++ front-end/src/app/_services/auth.service.ts
@@ -3,8 +3,7 @@ import { HttpClient } from '@angular/common/http';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { User } from '../_models/user';
 import { SessionStorageService } from './session-storage.service';
-import { JwtResponse } from '../_models/jwt-response';
-
+import decode from 'jwt-decode';
 @Injectable({
   providedIn: 'root'
 })
@@ -27,24 +26,23 @@ export class AuthService {
   } 
 
   deleteUser(userId: String): Observable<any> {
-    return this.httpClient.delete(this.AUTH_SERVER + "/api/user/" + userId);
+    return this.httpClient.delete(this.AUTH_SERVER + "/api/user/" + userId, {headers: {'Authorization':'Bearer ' + JSON.parse(this.getToken()).accessToken}});
   } 
 
   updateUser(userId: String, user: any): Observable<any> {
-    return this.httpClient.put(this.AUTH_SERVER + "/api/user/" + userId, user);
+    return this.httpClient.put(this.AUTH_SERVER + "/api/user/" + userId, user, {headers: {'Authorization':'Bearer ' + JSON.parse(this.getToken()).accessToken}});
   } 
 
   getAll() {
-    return this.httpClient.get<User[]>(this.AUTH_SERVER + "/api/user");
+    return this.httpClient.get<User[]>(this.AUTH_SERVER + "/api/user", {headers: {'Authorization':'Bearer ' + JSON.parse(this.getToken()).accessToken}});
   } 
 
   getByEmail(mail: String) {
-    return this.httpClient.get<User>(this.AUTH_SERVER + "/api/user/" + mail);
+    return this.httpClient.get<User>(this.AUTH_SERVER + "/api/user/" + mail, {headers: {'Authorization':'Bearer ' + JSON.parse(this.getToken()).accessToken}});
   } 
 
   logout() {
     this.isLogged.next(false);
-    this.sessionStorage.delete('user');
     this.sessionStorage.delete('accessToken');
   }
 
@@ -52,11 +50,15 @@ export class AuthService {
     return this.isLogged.value;
   }
 
-  getUser() {
-    return this.sessionStorage.get('user');
+  getToken() {
+    return this.sessionStorage.get('accessToken');
   }
 
-  setUser(user: any) {
-    this.sessionStorage.set('user', user);
+  getUser() {
+    return this.getToken() ? decode(JSON.parse(this.getToken()).accessToken) as User : null;
+  }
+
+  setToken(user: any) {
+    this.sessionStorage.set('accessToken', user);
   }
 }
